refactor: typescript化
Closes #1

+++ pages/api/session.ts
@@ -1,12 +1,17 @@
 import { sha512_256 } from 'js-sha512';
 import { timestampTotime } from '../../lib/util';
+import { NextApiRequest, NextApiResponse } from 'next';
 
-const { cert } = require('firebase-admin/app');
-const { getFirestore } = require('firebase-admin/firestore');
+import { cert } from 'firebase-admin/app';
+import { getFirestore } from 'firebase-admin/firestore';
 const serviceAccount = require('../../firebase'); //秘密鍵取得
-const admin = require('firebase-admin');
+import * as admin from 'firebase-admin';
+import { Session } from '@/types/session';
 
-export default async function handler(req, res) {
+export default async function handler(
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
   const COLLECTION_NAME = 'session';
   //　初期化
   if (admin.apps.length === 0) {
@@ -19,9 +24,9 @@ export default async function handler(req, res) {
 
   if (req.method === 'GET') {
     const snapshot = await db.collection(COLLECTION_NAME).get();
-    const slist = snapshot.docs.map((value, index, array) => {
+    const slist = snapshot.docs.map((value) => {
       return { id: value.id, ...value.data() };
-    });
+    }) as Session[];
     const nowd = new Date();
 
     //期限切れのセッションをデータベースから削除
@@ -33,12 +38,12 @@ export default async function handler(req, res) {
     });
 
     const cdata = slist.find((sdata) => sdata.id == req.query.id);
-    if (cdata && cdata.expiration > nowd) {
+    if (cdata && cdata.expiration && cdata.expiration > nowd) {
       res.status(200).json(cdata);
     } else {
       res.status(200).send('');
     }
-  } else if (req.method === 'DELETE') {
+  } else if (req.method === 'DELETE' && typeof req.query.id === 'string') {
     const sidh = sha512_256(req.query.id);
     await db.collection(COLLECTION_NAME).doc(sidh).delete();
     res.status(200).send('');
